# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventario API",
    "description": "REST API for Inventario — a multi-tenant, white-label asset management platform for sports federations, municipalities, schools, and clubs. Open-source under EUPL-1.2.",
    "version": "0.1.0",
    "contact": {
      "name": "Inventario · LTK Solutions",
      "url": "https://inventario.sportup.sk",
      "email": "inventario@sportup.sk"
    },
    "license": {
      "name": "EUPL-1.2",
      "url": "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token (v2.0). Multi-tenant: the JWT `tid` claim resolves the Organisation; the `oid` claim resolves the User. See apps/api/README.md for the device code flow."
      }
    },
...
  },
  "paths": {
...
    "/v1/locations": {
      "get": {
        "summary": "List locations",
        "tags": [
          "Locations"
        ],
        "description": "Returns a paginated list of locations, sorted by name. Soft-deleted locations are excluded. Optional filters: parentId, type, isActive.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "skip",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^([a-f\\d]{24}|null)$"
            },
            "in": "query",
            "name": "parentId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "WAREHOUSE",
                "OFFICE",
                "STADIUM",
                "TRAINING_CENTER",
                "EXTERNAL",
                "IN_TRANSIT"
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            },
            "in": "query",
            "name": "isActive",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "limit": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "skip": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "total",
                        "limit",
                        "skip",
                        "hasMore"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new location",
        "tags": [
          "Locations"
        ],
        "description": "Creates a new location. Slug must be unique (or derived from name if omitted). If parentId is supplied, the parent must exist and the resulting depth must not exceed the hierarchy limit. Requires ASSET_MANAGER or ADMIN role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 200
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "WAREHOUSE",
                      "OFFICE",
                      "STADIUM",
                      "TRAINING_CENTER",
                      "EXTERNAL",
                      "IN_TRANSIT"
                    ]
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "maxLength": 200
                      },
                      "city": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "postalCode": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "country": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2,
                        "default": "SK"
                      }
                    },
                    "additionalProperties": false,
                    "nullable": true,
                    "default": null
                  },
                  "coordinates": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": [
                      "lat",
                      "lng"
                    ],
                    "additionalProperties": false,
                    "nullable": true,
                    "default": null
                  },
                  "parentId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "nullable": true,
                    "default": null
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true,
                    "default": null
                  },
                  "managerId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "nullable": true,
                    "default": null
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "name",
                  "type"
                ],
                "additionalProperties": false,
                "description": "Telo pre vytvorenie lokality; slug je voliteľný (server odvodí z name)."
              }
            }
          },
          "description": "Telo pre vytvorenie lokality; slug je voliteľný (server odvodí z name)."
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/v1/locations/{id}": {
      "get": {
        "summary": "Get a single location by ID",
        "tags": [
          "Locations"
        ],
        "description": "Returns one location by its _id. 404 if not found or soft-deleted.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an existing location",
        "tags": [
          "Locations"
        ],
        "description": "Partial update — only provided fields are changed. Slug uniqueness, parent existence, and hierarchy invariants (cycle, depth) are revalidated on change. Records an audit event with a per-field diff. Requires ASSET_MANAGER or ADMIN role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 200
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "WAREHOUSE",
                      "OFFICE",
                      "STADIUM",
                      "TRAINING_CENTER",
                      "EXTERNAL",
                      "IN_TRANSIT"
                    ]
                  },
                  "address": {
                    "type": "object",
                    "properties": {
                      "street": {
                        "type": "string",
                        "maxLength": 200
                      },
                      "city": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "postalCode": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "country": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 2,
                        "default": "SK"
                      }
                    },
                    "additionalProperties": false,
                    "nullable": true,
                    "default": null
                  },
                  "coordinates": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                      },
                      "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                      }
                    },
                    "required": [
                      "lat",
                      "lng"
                    ],
                    "additionalProperties": false,
                    "nullable": true,
                    "default": null
                  },
                  "parentId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "description": "MongoDB ObjectId (24 hex znakov)",
                    "nullable": true,
                    "default": null
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true,
                    "default": null
                  },
                  "managerId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "description": "MongoDB ObjectId (24 hex znakov)",
                    "nullable": true,
                    "default": null
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "additionalProperties": false,
                "description": "Čiastočná aktualizácia lokality; všetky polia voliteľné."
              }
            }
          },
          "description": "Čiastočná aktualizácia lokality; všetky polia voliteľné."
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a location",
        "tags": [
          "Locations"
        ],
        "description": "Marks a location as deleted. Refuses deletion if the location has any non-deleted child locations (would orphan them), or if any non-deleted assets reference it. Requires ADMIN role.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    },
    {
      "url": "https://api.inventario.sportup.sk",
      "description": "Production (planned Q3 2026)"
    }
  ],
...
}